Webber-Attack/requirements.txt
# OpenRouter / LLM Integration
openai>=1.0.0

# MCP (Model Context Protocol)
mcp==1.25.0

# CLI + TUI
click==8.1.7
textual==0.47.1

# Database (MySQL)
mysql-connector-python==8.3.0

# HTTP (let pip resolve versions)
requests==2.31.0
httpx>=0.27.1
aiohttp==3.9.1

# Security Tools
python-nmap==0.7.1
dnspython==2.4.2
beautifulsoup4==4.12.2
python-whois==0.8.0
shodan==1.31.0
paramiko==3.4.0

# Integrations
PyGithub==2.1.1
gitpython==3.1.41
docker==7.0.0
ansible-core==2.16.3

# Reporting
jinja2==3.1.2
weasyprint==61.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
colorama==0.4.6
rich==13.7.0
pydantic>=2.11.0